requirements.txt
streamlit
requests
python-dotenv
Pillow
fpdf
deep-translator
langdetect